Implement efficient range-scans
Allow for the token-space to be split up into a configurable amount of
sub-ranges and concurrently scan a configurable subset of them. The
algorithm used is that described in Avi's efficent range scan blog post
[1].
The number of sub-ranges to split the token-space into can be set
by the `-range-count` command line paramaters. This defaults to 1, in
which case the table is scanned in a single query.
The concurrency can be set with the `-concurrency` command line
argument. Each sub-range will be an `op`.

[1] https://www.scylladb.com/2017/02/13/efficient-full-table-scans-with-scylla-1-6/