Data validation: use random data and checksum instead of simple pattern
The current data validation is quite simplistic and it depends on
knowing the size of the clustering row beforehand. This latter is
especially problematic since we want introduce support for random
clustering row sizes.
To prepare for this, data generation and validation is refactored.
The data blob now consists of:
* The size of the blob;
* The primary key;
* The clustering key;
* A random payload;
* A sha256 checksum calculated over the payload.

This format allows for easily validating that:
* The blob was not truncated or extended (stored size == actual size);
* The blob comes from the partition and row it is expected to come from;
* The blob was not changed. The checksum should be able to detect even a
  single bit flipped, while collisions (two rows having the same data)
  is practically impossible.

It should be noted that only this "full format" is only possible to use
when the size of the row is greater than a certain size (57 bytes,
all mandatory fields + a single byte payload). When the size of the
message is below this a special, deterministic format is used that has
reduced capabilities of detecting corruption and avoiding collisions.

It would also be quite simple to add a generation field and thus detect
resurrected data.