Add infrastructure for using random value distributions
This patch adds the infrastructure for using random value distributions
for e.g. clustering row sizes.
It uses the existing random package to provide the random distributions.

The command-line interface remains backward compatible. If a single
number is passed, a fixed distribution is created.

Also add a struct implementing the `flag.Value` interface, so distribution
values can be easily used as command-line arguments.