Print time and latency with reasonable precision
When the current code prints times and latencies, it uses the default way
of printing a time.Duration, which prints absurdly precise numbers, and
the following example output shows:

 time                operations/s   rows/s   errors  max              99.9th           99th             95th             90th             median           mean
 9.006293118s               26534    26534        0  3.047423ms       1.966079ms       1.179647ms       917.503µs        819.199µs        589.823µs        602.173µs

The very long lines are inconvenient to view, and also not helpfu
(nobody needs to print whole seconds in nanosecond (!) precision, or
to add 3 more digits of precision to 589 microseconds).

This patch adds a function to Round() time.Duration objects for printing,
and uses it. The result is now more useful, clearer, and easily fits
in a window with a line width of 80:

 time   ops/s  rows/s errors max    99.9th 99th   95th   90th   median mean
   1s   27517   27517      0 2.7ms  1.7ms  1.1ms  885µs  786µs  590µs  581µs

Fixes #44

Signed-off-by: Nadav Har'El <nyh@scylladb.com>