readme: document the correct way to build shard-aware scylla-bench
Currently, README.md suggests using the `go get` command to download and
install scylla-bench. However, this method results in a scylla-bench
binary being built with the upstream version of the gocql driver, not
the ScyllaDB's fork - and the latter is required for the shard-awareness
to work.

Considering that scylla-bench was created for benchmarking Scylla, it
makes more sense to build it with our fork of gocql. This commit
improves the README.md: a correct way to build scylla-bench is now
included with an explanation why simply using `go get` does not work.