Reduce err msg size for `batch write` queries
When a 'batch write' command fails we print error message that
contains 'ck' and 'value size' values for each of the queries
used in that batch.
And if we have really huge number of rows per request then we get
really huge err message.
It is useless and, moreover, greatly increases logs size.

Fix it by:
- printing 'ck range' like '9410000..9411999' instead of printing
  all the values (we have configurations when it is 2000)
  which always differ by 1 with neighbours.
- printing 'avg values size' instead of all of them.

Above message fixes make a batch error message be predictable enough
and not dependent on the 'rows per query' configuration value.