fix(host-verification): switch to gocql builtin hostname verification
Custom logic of providing server name as a separate scylla-bench parameter,
for host verification, was introduced a few years ago.
But now host verification is performed by default by the driver itself, when TLS
encryption is enabled and gocql.SslOptions.EnableHostVerification flag is set.

The change switches to default host verification mechanism and removes
the corresponding customization from the scylla-bench.

Fixes: https://github.com/scylladb/scylla-bench/issues/140