Fix row-thread mapping for sequential workloads
Before this change scylla-bench was distributing rows among threads
correctly only when number of partitions was divisible by number of
threads without remainder.

A bit worse case was appearing when there was division remainder.

And the worst case was when number of threads was bigger than number
of partitions. In this case last thread was handling all of the rows
and all of other threads were idle.

So, fix it by spreading rows evenly among all of the threads
as much as possible.

Closes: #138