fix(release): make extract-driver-version.sh executable
A helper to extract scyla-gocql driver version info during CI release
workflow was introduced in
https://github.com/scylladb/scylla-bench/pull/179.
Currently the script is made executable during the CI release workflow
itself, which is not tolerated by the GoReleaser as it detects dirty
git state and fails the workflow.

The fix makes the extract-driver-version.sh executable out-of-the-box in the repo.