hdr: it reports endtime instead of interval
There is a bug in HdrHistogram/hdrhistogram-go that makes it report
endtime instead of interval.
And another bug that makes it report interval and start time in
milliseconds.

Both these problems addressed in my fork: dkropachev/hdrhistogram-go
PR created to port it to the upsteam: https://github.com/HdrHistogram/hdrhistogram-go/pull/51